Jumbled Word Solver
Finds the possible permutations of a Jumbled word and cross-references
with a dictionary to filter out which permutations are actual English
words.
Permutations are written to Results.txt
and the results are compared with the list of 355k+ words in
Dictionary.txt.
Depending on your processor speed, the time taken to finish might differ
from PC to PC. So I've added a feature that plays a sound when
everything is complete.